+++ src/pedido/producto-negocio.ts
@@ -1,0 +1,36 @@
+import { Injectable } from "@nestjs/common";
+import { CrearPedidoDTO } from "./input-dtos/crear-pedido-dto";
+import { Pedido } from "src/entities";
+import { PedidoService } from "./pedido.service";
+import { NegocioException } from "src/utils/exceptions/negocio-exception";
+
+@Injectable()
+export class ProductoNegocio {
+
+    constructor(private readonly pedidoService: PedidoService) { }
+
+    async crearPedido(pedido: CrearPedidoDTO): Promise<Pedido> {
+        this.crearPedidoValidaciones(pedido);
+        return this.pedidoService.crearPedido(pedido);
+    }
+
+    private crearPedidoValidaciones(pedido: CrearPedidoDTO) {
+        if (!pedido.tiendaId || pedido.tiendaId.length === 0) {
+            throw new NegocioException('Por favor, selecciona una tienda para atender este pedido.');
+        }
+        if (!pedido.paquetes || pedido.paquetes.length === 0) {
+            throw new NegocioException('Debe haber al menos un paquete en el pedido.');
+        }
+        if(pedido.paquetes || pedido.paquetes.length > 0){
+            pedido.paquetes.some(paquete => {
+                if (!paquete.productoId) {
+                    throw new NegocioException('Debe seleccionar algún producto para el paquete.');
+                }
+                if (paquete.cantidad < 1) {
+                    throw new NegocioException('Los paquetes del pedido deben contener al menos una unidad.');
+                }
+            })
+        }
+    }
+    
+}
